CareerFinance-AI_V3-main - Copy/backend/requirements_rag.txt
# Dépendances pour le système RAG LinkedIn
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0

# Dépendances existantes (à vérifier si déjà installées)
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
pdfplumber
pytesseract
pillow
pydantic
